Improved bookmarks support in PathParameterValueWidget.
It now finds the appropriate application bookmarks as long as either the window it is in or the plug it represents eventually lead back to an application in some way. This enables bookmarks for parameters in the floating OpDialogues created in the browser app. This implements the most important part of #787.