Added per-column editability to CompoundVectorParameterValueWidget.
This uses a ["UI"]["editable"] user data entry in each child parameter, where a BoolData value of False will make the column for that parameter read-only. I would have liked to have used the existing ["gaffer"]["readOnly"] entry which controls the editability of the underlying plug values instead, but that would prevent resizing of the data when only some columns are editable.

Fixes #766.