Made grid width minimum > 0 to avoid OpenGL errors.
As far as I understand it, the GL spec specifies that a width of 0 may be specified, but will be clamped up to the next representable value. That doesn't seem to be the case with either the OS X or NVIDIA drivers though, where a width of 0 causes errors.

Also enabled line smoothing for the grid, and changed the default values to take better advantage of the new quality this provides.