Added command-line representation of Op values in the UI.
This is implemented by providing a "summary" Metadata entry for the node which contains the command line, and then using a small _InfoButton() class to display it. The summary is displayed both as a tooltip and also in a popup window when the button is clicked.

Since the NodeGraph also implements tooltips using the "summary" Metadata, the summary would also be available there if we used Ops there. We should implement summaries for more node types and then consider reusing the InfoButton more generally.

Fixes #793.