Improved support for CompoundVectorParameter column editability.
The user data specifying per-column editability can now be changed from a parameterChanged() callback, and will be reflected in the UI. Note that this will currently only work if the parameterChanged() callback was triggered by a parameter within the CompoundVectorParameter itself, as changes to other parameters will not trigger the _updateFromPlug() method needed to apply the changes. We may be able to improve on this in the future if we map all parameter user data into per-plug metadata entries and have the UI respond to changes in that, but that is a bigger task.