Added an auto-load preset for ops.
This is created using the usual preset save dialogue, and loaded automatically by the op app and the browser app in op mode. The preset is not auto-loaded in the case that the op app command line specifies a different preset or any arguments - in that case it is assumed that the caller wishes to specify the parameter values themselves.

Fixes #804, and hopefully doesn't cause any "your pipeline is now entirely at the mercy of presets anyone can make at any time" problems of its own.