Fixed bug in ImageTransform.
This was caused by ImageTransform::hashChannelData() not correctly including the tile origin in the hash. For transforms where two output tiles share the exact same set of input tiles affecting them, we were computing exactly the same hash for both output tiles, even though they each would filter the pixels in the input tiles totally differently.

This bug was introduced in 78b8739f1867ecd3b660838d384cdd6083734d92, where I refactored the hash computation for images. Specifically, in that commit it became the responsibility of derived classes to hash in the tile origin where necessary whereas before the SceneProcessor _always_ hashed in the tile origin whether it was necessary or not. The old code resulted in unnecessary computation for networks like Constant->Grade, where although all the tiles have the same hash in the Constant, they would have different hashes in the Grade. The new way of doing things is for sure better, provided you're not a numpty like I was and fail to hash in the tile origin where relevant.

Also renamed the test case, so it actually makes sense now we've fixed things - in any case we should have used @GafferTest.expectedFailure to decorate it in the first place so we could still give it a meaningful name.