Simplified ImageTransform::channelDataHash() and removed todo.
Using the less verbose method of hashing all the children of a CompoundPlug.

The comment in the todo was inaccurate - ImageProcessor no longer does any hashing of the tile origin. We also won't be changing the coordinate space that tile origin is specified in - tile origin, display window, and data window are all specified relative to the same origin for good reasons - sanity and ease of understanding. There's no reason the described optimisation couldn't work with the origin we have, but it seems best to implement in a node which specifically does axis-aligned integer shifts in the image without any filtering.