Optimised PathMatcher.
Added new methods for adding and removing paths as vector<InternedString> - this avoids all the overhead of tokenization for cases where the path already exists in tokenized form.

Changed internal representation to use InternedString rather than std::string. This is much more compact than the old representation, since most strings we put into a PathMatcher will already have an InternedString representation somewhere (from the SceneNodes or from IndexedIO).

PathMatcherTest.testLookupScaling() shows a 67% reduction in time to build the PathMatcher, a 78% reduction in time to perform the lookups, and a 70% reduction in memory usage.