Adjusted SceneProceduralTest for new SceneReader behaviour.
For the test, we need a node which will throw during procedural expansion. The SceneReader no longer fits that bill now that it implements computeGlobals(), as it will now throw during construction of the SceneProcedural. Using the Text node with an invalid font gives us an exception at the right time.