Added pausing of interactive renders.
Previously pausing a render just meant that we stopped sending edits, but the render was still happening in the background and taking up valuable cycles. We now pause the renderer itself in addition to not sending edits any more.

Refactored the code a bit to make it all easier to follow. Before, we were changing our internal state and stopping/starting renders in various different signal handlers and it was hard to follow the overall logic. Now we use much simpler signal handlers which just track what has been dirtied, and keep all our update logic in a single update() method, which is hopefully easier to follow.

Fixes #646.