Fixed Viewer shutdown error.
Now we have the lifetime of the ScriptNode sorted out properly, the Viewer was emitted the following message on quit every time :

`ReferenceError: Instance referenced by WeakMethod __updateRequest() no longer exists`

This turned out to be because there was a View->Toolbar->View circular reference keeping the View alive past the lifetime of the Viewer. Thus the Viewer was requesting updates but the Viewer wasn't there to service them. Removing the circular reference fixes the problem.