Stopped ApplicationMenu creating zombie widgets.
It used to store the About window and Preferences windows as globals, meaning they could never ever die. They are now stored on the ScriptWindows they get parented to, and only weakrefs are used in the globals to help track them. In this case, no real harm was done, but since other UIs can hold on to significant resources (and cause shutdown crashes if they hold InteractiveRender nodes at shutdown), it's worth being strict everywhere in order to be able to warn of genuine problems at shutdown.