Added public Filter methods for specifying input scene via Context.
Prior to this, it wasn't possible to query a filter from Python any more. The new methods also mean that FilteredSceneProcessor doesn't need to know anything about how filters are implemented (which is good, because it's a horrible dirty secret), and we hide some more implementation by making the name of the context entry for the input scene private.