Added _copy parameter to Shader::state() python binding.
This follows the same rationale as TypedObjectPlug::getValue() - boost::python insists on a non-const return value, but we don't trust python code not to modify the data, so we return a copy as a matter of course. The additional parameter allows the original data to be returned, and is intended only really for use in test cases. The current motivation is to be able to accurately benchmark shader performance from python without having the results marred by unnecessary overhead.