ExecutableRender::executionHash now accounts for the input node and the context.
Note that we are hashing the address of the input node to work around the lack of a hierarchyHash. This should be safe because executionHash needs to be valid during Dispatcher::dispatch(), and no nodes should be deleted during that time.

I haven't added tests for ArnoldRender because I can't run it locally, but the OpenGL and RenderMan tests are nearly identical, and ArnoldRender doesn't override executionHash.