Parallelised Instancer::hashBranchTransform().
Because the computation of bounds generally occurs before traversing to the children, the hashing of the union of the bounds of all the instances actually represents a serial bottleneck, which can be quite large when there are a large number of instances. Parallelising this operation reduces the bottleneck, and stops cores sitting idly waiting for work to do.

This gives ~1.8x speedup on a 2 core laptop and ~5.6x speedup on a 6 core workstation.