Implemented update-on-dirty for SceneInspector inheritance popup windows.
The SceneInspector class can now be instantiated with a specific set of Sections, and the target paths may be pinned. The inheritance diagram is now implemented as a section, meaning we can simply put a small SceneInspector in the popup window and take advantage of all the update machinery implemented there. We could also use the same mechanism to tear off copies of any SceneInspector section if we wanted to in the future.