Fixed SceneInspector inheritance window layout glitch.
I don't fully understand what is happening but I think it's because when we use the auto-parenting mechanism for a parent which is already visible, Qt realises the child QWidgets earlier in the process, and then fails to update the layout when we change their properties subsequently. By using `_mainColumn()[:] = rows` we not only avoid this, but also do the deletion of existing rows and addition of new ones in one go, which cuts down on intermediate updates and gives a snappier build.