Made EventLoop remove idle callbacks immediately after calling.
When an idle callback was to be removed (because it returned False), it wasn't removed immediately, but only after calling all other idle callbacks. This made it impossible to re-add the removed callback from one of those subsequent callbacks (an assertion was fired). The removal is now immediate so the callback can be re-added if necessary.