Added hotspots to Pointer class.
This allows the center of the pointer to be specified, so there are no longer jumps when switching between pointers. Pointer instances can now be instantiated, and these store an image along with the position of the hotspot. Added new setCurrent()/getCurrent() methods for specifying which instance is the current one, and a registry for storing a map of named Pointer types which can be used throughout the rest of the code.

The old methods remain for now, but all use of them has been removed.