Fixed SceneInspector transform source labelling.
It was mistakenly setting one corner widget twice rather than setting both once. I discovered this because it triggered a horrendous crash bug in PySide (but not PyQt, which still seems far the stabler of the two).

For future reference, a distilled reproduction for the original bug can be found here :

https://gist.github.com/johnhaddon/de48fbde0031017cdc03