Renamed GafferScene.Displays node to GafferScene.Outputs.
General consensus is that this is more descriptive of its purpose, and will also reduce Displays/Display confusion. However, this does violate one of the basic principles in Gaffer - that naming should be consistent throughout the UI, scripting layer and underlying libraries. The intention is to address this in future by renaming the IECore.Display class to IECore.Output. The rationale for doing the Gaffer rename on its own now is that this is our last chance to change it before we're relying on this particular node in production.

Fixes #54.