Added global mode to GafferScene::Attributes node.
This places the attributes in the globals (with "attribute:" prefixes) rather than at locations in the scene. This is for use in setting up "render globals" networks where in fact some of the values being specified are actually renderer attributes rather than options (maximum ray depth in 3delight for instance).