Added Context::changed() method.
This is an official means of letting the Context know that an entry has been changed behind its back (through shared or borrowed ownership for instance). It avoids constructing the changed signal unnecessarily (previously it was documented that such changes must be accompanied by a manual emission of the signal), and ensures that the lazy hash computation is invalidated appropriately.