Made Context::hash() ignore "ui:" prefixed entries.
In general these entries have become a bit of a nuisance - there are now many places where we have to ignore them because they're not relevant to computation, and we don't want small changes in the ui to trigger recomputes. There's a good case for making the UI use a second structure for UI specific entries, but for now we do the pragmatic thing and ignore them in one more place. On the bright side, we can now remove special case code from ExecutableRender and SceneWriter, so the net effect is one less special case.