Fixed crashes when passing None to PathMatcher python methods.
Boost python will happily extract a NULL pointer from None when asked for any pointer type, so we need to guard against it. In this case we can replace some wrapping functions and allow the implicit ScenePath conversion registered in ScenePlugBinding.cpp to do the work (this was added after the offending code).