Fixed Display node for bucket sizes larger than the native tile size.
We were incorrectly applying the x offset of the transfer area within the bucket data, by not taking into account the number of channels (the source channels are interleaved). This hadn't been seen noticed till now because we'd never rendered with a 3delight bucket size greater than the Gaffer tile size.