FrameRange UI displays the value that will actually be dispatched.
This overrides the widget text on the Dispatcher frameRange plug to show the value that will be dispatched, even when not in CustomRange mode. The actual value of the plug remains the same, but the widget displays the CurrentFrame or FullRange when in those modes, and the PlaybackRange updates automatically in that mode (which does change the plug value).