Moved Cortex-specific functionality into new GafferCortex library and module.
The Gaffer core itself remains heavily dependent on Cortex, and always will, but Cortex is a huge beast, and it doesn't make sense to put everything Cortex related into the core of Gaffer. Here we're splitting out "end user" functionality such as OpHolders and ProceduralHolders, so that GafferCortex can be thought of as the user-visible presence of Cortex within Gaffer. This leaves behind a smaller libGaffer where the core concepts of Nodes and Plugs aren't confused by  non-core concepts like Ops and Parameters.

This commit doesn't move the test cases into GafferCortexTest, but instead leaves them in GafferTest with minor tweaks to accept the new TypeIds, but not to use the new namespace. The tests still pass due to the presence of the gafferCortexCompatibility.py startup file, which allows a grace period for dependent code to update to the new module layout.