Added child matching to connected Plugs.
When a parent plug has an input connection from another parent plug, we now propagate any child addition/removal on the source plug, performing equivalent operations on the destination plug. This allows the connection to be maintained, and will allow compound plug types like CompoundDataPlug to be used effectively in connections.