Fixed context errors in the scene inspector
We noticed a problem in a scene where an expression was trying to pull a variable from the context. This variable was defined in script['variables'], and should have been available to the SceneInspector, but wasn't as its __update method was being called with an incorrect context. This was causing errors while trying to use the SceneInspector. I've added a 'with self.getContext()' in SceneInspector.__update() to address this.