Fixed serialization bug with extra plugs on references
The reference nodes we use in production have an extra "__jabuka" CompoundPlug on them, with a bunch of children containing asset info. The children's names aren't prefixed by a double underscore, so Reference::isReferencePlug thinks they're referenced, and wipes out their Plug::Dynamic flag when we reload a reference. This means the node's asset tracking info gets destoyed when the reference gets reloaded and we load/save the file. I've fixed this by classifying descendants of a CompoundPlug with a double underscore in its name as non referenced.