Added GafferSceneUI::SceneGadget class.
This contains a small OpenGL scene graph implementation which can be selectively updated based on plug dirtiness in Gaffer. This is an improvement over using IECoreGL::Renderer with GafferScene::SceneProcedural, which has to regenerate the entire scene for every Gaffer change.

This is intended to form the basis for scene viewer with improved performance.