Fixed SceneView for compilation with GCC 4.1.2.
Because TypedData specialisations must happen in the IECore namespace, we define PathMatcherData in IECore and then typedef it into GafferScene. GCC 4.1.2 then considers PathMatcherData to be an ambiguous name if we have "using namespace IECore" and "using namespace GafferScene". Other compilers realise that both refer to the same name anyway and just get on with things.