Fixed bug in Instancer dirty propagation.
This had gone unnoticed before because the viewer would always do a full update when any plug was dirtied. Now that updates are done partially based on the specific dirty plugs, it is much more critical that SceneNode::affects() implementations are accurate. We'll need to be on the lookout for similar bugs as we adopt the new SceneView.