Fixed bug in StandardGraphLayout node insertion code.
The bug appears to have been exposed by the introduction of the internal pass-through connections in the recent GafferScene optimisations. The test necessary to skip the target node when considering insertion targets existed only in the "make the connections" loop, and not in the "can we make the connections loop?". Moving all the tests into only one loop avoids duplicating them and helps ensure similar bugs won't creep in later on.