Optimised SceneGadget for static geometry.
When the hash for the object is the same as before, we can entirely skip the process of computing the object in Gaffer and then reconverting it to GL. This avoids a lot of cache churn in the presence of unchanging objects - for instance when switching between nodes with identical hierarchies and objects, or when playing back an animation with a large number of static objects. Processing of 25 frames of the bunny test scene for GL display is approximately 6 times faster with this change.